seperated files and added more detailed code + hashed mac address
